refactor: move deployment helpers to drivers/common
These modules are only used directly by drivers, so they should live
with the drivers.
