+++ lib/drivers/common/pack-receiver.js
