+++ lib/drivers/common/prepare.js
@@ -2,7 +2,7 @@ var _ = require('lodash');
 var async = require('async');
 var debug = require('debug')('strong-pm:prepare');
 var npm = require('strong-spawn-npm');
-var mandatory = require('./util').mandatory;
+var mandatory = require('../../util').mandatory;
 
 // Private properties are for tests.
 exports._COMMANDS = COMMANDS;
