+++ lib/drivers/direct/container.js
@@ -9,11 +9,11 @@ var c2s = require('strong-runner').Runnable.toString;
 var cicada = require('strong-fork-cicada');
 var cicadaCommit = require('strong-fork-cicada/lib/commit');
 var fs = require('fs');
-var localDeploy = require('./../../local-deploy');
+var localDeploy = require('../common/local-deploy');
 var mandatory = require('./../../util').mandatory;
-var packReceiver = require('./../../pack-receiver');
+var packReceiver = require('../common/pack-receiver');
 var path = require('path');
-var prepare = require('./../../prepare').prepare;
+var prepare = require('../common/prepare').prepare;
 var rmrf = require('rimraf');
 var util = require('util');
